fix: retrieve activities from strava

+++ fivekmrun_app_flutter/lib/common/double_extensions.dart
@@ -1,4 +1,4 @@
-extension DoubleExtension on int {
+extension DoubleExtension on double {
   String metersToKm() {
     return (this / 1000).toStringAsFixed(2);
   }
